Rename folder to frontier/ so it is consistent with namespace
