+++ lib/frontier/authorization.rb
