+++ lib/frontier/authorization/assertion.rb
