+++ lib/frontier/authorization/check.rb
