+++ lib/frontier/authorization/scope.rb
