+++ lib/frontier/hash_decorator.rb
