+++ lib/frontier/input.rb
