+++ lib/frontier/input/association.rb
