+++ lib/frontier/input/attribute.rb
