+++ lib/frontier/input/factory.rb
