+++ lib/frontier/migration_string_builder.rb
