+++ spec/lib/frontier/authorization/assertion_spec.rb
