+++ spec/lib/frontier/authorization/check_spec.rb
