+++ spec/lib/frontier/authorization/scope_spec.rb
