+++ spec/lib/frontier/hash_decorator_spec.rb
