+++ spec/lib/frontier/input/association_spec.rb
