+++ spec/lib/frontier/input/attribute_spec.rb
