+++ spec/lib/frontier/input/factory_spec.rb
