+++ spec/lib/frontier/migration_string_builder_spec.rb
